+++ src/ts/Native/NativeBridge.ts
@@ -21,6 +21,9 @@ import { IosAdUnitApi } from 'Native/Api/IosAdUnit';
 import { NotificationApi } from 'Native/Api/Notification';
 import { UrlSchemeApi } from 'Native/Api/UrlScheme';
 import { LifecycleApi } from 'Native/Api/Lifecycle';
+import { AndroidPreferencesApi } from 'Native/Api/AndroidPreferences';
+import { IosPreferencesApi } from 'Native/Api/IosPreferences';
+import { SensorInfoApi } from 'Native/Api/SensorInfo';
 
 export enum CallbackStatus {
     OK,
@@ -46,12 +49,14 @@ export class NativeBridge implements INativeBridge {
 
     public AppSheet: AppSheetApi;
     public AndroidAdUnit: AndroidAdUnitApi;
+    public AndroidPreferences: AndroidPreferencesApi;
     public Broadcast: BroadcastApi;
     public Cache: CacheApi;
     public Connectivity: ConnectivityApi;
     public DeviceInfo: DeviceInfoApi;
     public Intent: IntentApi;
     public IosAdUnit: IosAdUnitApi;
+    public IosPreferences: IosPreferencesApi;
     public Listener: ListenerApi;
     public Lifecycle: LifecycleApi;
     public Notification: NotificationApi;
@@ -59,6 +64,7 @@ export class NativeBridge implements INativeBridge {
     public Request: RequestApi;
     public Resolve: ResolveApi;
     public Sdk: SdkApi;
+    public SensorInfo: SensorInfoApi;
     public Storage: StorageApi;
     public VideoPlayer: VideoPlayerApi;
     public UrlScheme: UrlSchemeApi;
@@ -84,8 +90,10 @@ export class NativeBridge implements INativeBridge {
 
         if(platform === Platform.IOS) {
             this.IosAdUnit = new IosAdUnitApi(this);
+            this.IosPreferences = new IosPreferencesApi(this);
         } else {
             this.AndroidAdUnit = new AndroidAdUnitApi(this);
+            this.AndroidPreferences = new AndroidPreferencesApi(this);
         }
 
         this.Broadcast = new BroadcastApi(this);
@@ -100,6 +108,7 @@ export class NativeBridge implements INativeBridge {
         this.Request = new RequestApi(this);
         this.Resolve = new ResolveApi(this);
         this.Sdk = new SdkApi(this);
+        this.SensorInfo = new SensorInfoApi(this);
         this.Storage = new StorageApi(this);
         this.VideoPlayer = new VideoPlayerApi(this);
         this.UrlScheme = new UrlSchemeApi(this);